added introScene as a component
