+++ src/components/chapter2/start/scene/sceneComponents/blackCut.tsx
@@ -1,0 +1,9 @@
+import React from "react";
+import styles from "../../start.module.css";
+
+export const BlackCut = () => {
+  return (
+    <div className={`${styles.blackCut} w-full absolute`}
+         style={{ background: "black", height: document.documentElement.clientHeight }}>
+    </div>);
+};
