+++ src/components/chapter2/start/scene/sceneComponents/car.tsx
@@ -1,21 +1,14 @@
 import React from "react";
 
-import styles from "../start.module.css";
-import img from "../asset/kapitel2_fahrrad_auswahl.png";
-import bycicleImg from "../asset/kapitel2_fahrrad_auswahl_effekt.png";
+import styles from "../../start.module.css";
+import img from "../../asset/kapitel2_auto_auswahl.png";
+import carImg from "../../asset/kapitel2_auto_auswahl_effekt.png";
 import Plx from "react-plx";
-import { Sequence } from "../interface/Chapter2Enum";
-import { Selection } from "../../../topic-selection";
-import { TopicSelection } from "../../../topic-selection/interfaces/selection";
-import { Quizinterface } from "../../../quiz-statistics/interface/quiz-interface";
-import { getParallaxData } from "../animationParallaxData";
-import { getContentData, getStatsData } from "../content";
-import { getQuiz } from "../quizData";
+import { Sequence } from "../../interface/Chapter2Enum";
+import { Selection } from "../../../../topic-selection";
+import { getParallaxData } from "../../animationParallaxData";
 
-import { Stats } from "../interface/interface";
-import carImg from "../asset/kapitel2_auto_auswahl_effekt.png";
-
-export const Bicycle: React.FC<{
+export const Car: React.FC<{
   parallaxData: [];
   sequence: Sequence;
   setSequence: React.Dispatch<React.SetStateAction<Sequence>>;
@@ -30,22 +23,22 @@ export const Bicycle: React.FC<{
     setViewHeight("2000vh");
     setParallaxData(
       getParallaxData(
-        Sequence.Bicycle,
+        Sequence.Car,
         document.documentElement.clientHeight,
         document.documentElement.clientWidth
       )
     );
-    setSequence(Sequence.Bicycle);
+    setSequence(Sequence.Car);
   };
 
   const topicSelection = sequence === Sequence.Question ?
     {
       containerStyle: { opacity: silloutteOpacity },
-      containerClassName: styles.bicycleChoice,
+      containerClassName: styles.carChoice,
       selection: [
         {
           style: { width: "100%" },
-          imagePath: bycicleImg,
+          imagePath: carImg,
           mouseEnterEventHandler: mouseEnterHandlerFunction,
           mouseLeaveEventHandler: mouseLeaveHandlerFunction,
           mouseClickEventHandler: mouseClickEventHandlerFunction
@@ -54,9 +47,9 @@ export const Bicycle: React.FC<{
     } : undefined;
 
   return (
-    <div id="bicycle" className={`${styles.bicycle} absolute -top-96`}>
+    <div className={`${styles.car} absolute -top-96 `}>
       <Plx parallaxData={parallaxData} animateWhenNotInViewport={true}>
-        <img src={img} alt="Bicycle" className="absolute" />
+        <img id="car" src={img} alt="car" className="absolute" />
         {topicSelection !== undefined ? (
           <Selection topicSelection={topicSelection} />
         ) : null}
