+++ src/components/chapter2/start/start.module.css
@@ -93,6 +93,11 @@
         zIndex: 2;
     }
 
+    .blackCut {
+        top: 58rem;
+    }
+
+
     .graffiti {
         top: 0rem;
     }
